always use Union as a state
