+++ src/main/java/com/yahoo/sketches/hive/hll/DataToSketchUDAF.java
@@ -99,22 +99,6 @@ public class DataToSketchUDAF extends AbstractGenericUDAFResolver {
 
   public static class DataToSketchEvaluator extends SketchEvaluator {
 
-    private Mode mode_;
-
-    @SuppressWarnings("deprecation")
-    @Override
-    public AggregationBuffer getNewAggregationBuffer() throws HiveException {
-      // Different State is used for the iterate phase and the merge phase.
-      // A user reported that in some version of Hive this was apparently called before init,
-      // so the mode_ was null. A solution was implemented to have UnionState, which can work
-      // in both cases, but SketchState is more space-efficient.
-      // HLL sketch is about compactness, so let's use SketchState if possible.
-      if ((mode_ == Mode.PARTIAL1) || (mode_ == Mode.COMPLETE)) { // iterate() will be used
-        return new SketchState();
-      }
-      return new UnionState();
-    }
-
     /*
      * (non-Javadoc)
      *
@@ -126,7 +110,6 @@ public class DataToSketchUDAF extends AbstractGenericUDAFResolver {
     @Override
     public ObjectInspector init(final Mode mode, final ObjectInspector[] parameters) throws HiveException {
       super.init(mode, parameters);
-      mode_ = mode;
       if ((mode == Mode.PARTIAL1) || (mode == Mode.COMPLETE)) {
         // input is original data
         inputInspector_ = (PrimitiveObjectInspector) parameters[0];
